requirements.txt
certifi==2020.6.20
click==8.0.3
dataclasses==0.8
Flask==2.0.2
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
paho-mqtt==1.6.1
typing_extensions==4.0.1
Werkzeug==2.0.2
zipp==3.6.0
